pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workday-cli"
version = "0.1.0"
description = "Command-line pomodoro timer with Telegram notifications"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "skvost"}
]
keywords = ["pomodoro", "timer", "productivity", "cli", "telegram"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
]

dependencies = [
    "click>=8.0",
    "python-telegram-bot>=20.0",
    "toml>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
workday = "workday.cli:main"
wd = "workday.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["workday*"]
